+++ graphdatascience/query_runner/progress/progress_provider.py
@@ -8,6 +8,7 @@ class TaskWithProgress:
     task_name: str
     progress_percent: str
     status: str
+    sub_tasks_description: Optional[str] = None
 
 
 class ProgressProvider(ABC):