+++ graphdatascience/query_runner/progress/query_progress_provider.py
@@ -21,16 +21,33 @@ class QueryProgressProvider(ProgressProvider):
 
     def root_task_with_progress(self, job_id: str, database: Optional[str] = None) -> TaskWithProgress:
         tier = "beta." if self._server_version_func() < ServerVersion(2, 5, 0) else ""
-        # we only retrieve the progress of the root task
+
+        # expect at exactly one row (query will fail if not existing)
         progress = self._run_cypher_func(
             f"CALL gds.{tier}listProgress('{job_id}')"
             + " YIELD taskName, progress, status"
-            + " RETURN taskName, progress, status"
-            + " LIMIT 1",
+            + " RETURN taskName, progress, status",
             database,
-        ).squeeze()  # expect at exactly one row (query will fail if not existing)
+        )
 
-        progress_percent = progress["progress"]
-        root_task_name = progress["taskName"].split("|--")[-1][1:]
+        # compute depth of each subtask
+        progress["trimmedName"] = progress["taskName"].str.lstrip()
+        progress["depth"] = progress["taskName"].str.len() - progress["trimmedName"].str.len()
+        progress.sort_values("depth", ascending=True, inplace=True)
 
-        return TaskWithProgress(root_task_name, progress_percent, progress["status"])
+        root_task = progress.iloc[0]
+        root_progress_percent = root_task["progress"]
+        root_task_name = root_task["trimmedName"].replace("|--", "")
+        root_status = root_task["status"]
+
+        subtask_descriptions = None
+        running_tasks = progress[progress["status"] == "RUNNING"]
+        if running_tasks["taskName"].size > 1:  # at least one subtask
+            subtasks = running_tasks[1:]  # remove root task
+            subtask_descriptions = "::".join(
+                list(subtasks["taskName"].apply(lambda name: name.split("|--")[-1].strip()))
+            )
+
+        return TaskWithProgress(
+            root_task_name, root_progress_percent, root_status, sub_tasks_description=subtask_descriptions
+        )